magna-ai-backend/backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.6.0
sse-starlette==2.1.3

# LLM Providers
google-generativeai==0.8.3
openai==1.57.0
langchain==0.3.13
langchain-community==0.3.13
litellm==1.52.14

# Vector Database and Embeddings
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
pinecone-client==5.0.1
numpy>=1.25.0

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.10

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Caching
cachetools==5.3.2

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==44.0.0

# AWS S3
boto3==1.35.76
botocore==1.35.76

# Search
serpapi==0.1.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
hypothesis==6.122.3
faker==33.1.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2

# Monitoring and Logging
structlog==24.4.0
python-json-logger==3.2.1
